Fix `size_t` related issues by cast to `u32`

+++ src/xrCore/LocatorAPI.cpp
@@ -429,7 +429,7 @@ void CLocatorAPI::ProcessArchive(LPCSTR _path)
 
 	m_archives.push_back		(archive());
 	archive& A					= m_archives.back();
-	A.vfs_idx					= m_archives.size()-1;
+	A.vfs_idx					= (u32)m_archives.size()-1;
 	A.path						= path;
 
 	A.open						();
